requirements.txt
opencv-python
numpy
Pillow

# gradio_client
websocket-client